--- src/pages/summer/Main/Footer/Footer.js
@@ -1,47 +1,0 @@
-import React from 'react';
-
-import './Footer.scss';
-
-class Footer extends React.Component {
-  render() {
-    return (
-      <footer>
-        <ul>
-          <li>
-            <a href="#">소개</a>
-          </li>
-          <li>
-            <a href="#">블로그</a>
-          </li>
-          <li>
-            <a href="#">채용 정보</a>
-          </li>
-          <li>
-            <a href="#">도움말</a>
-          </li>
-          <li>
-            <a href="#">API</a>
-          </li>
-          <li>
-            <a href="#">개인정보처리방침</a>
-          </li>
-          <li>
-            <a href="#">약관</a>
-          </li>
-          <li>
-            <a href="#">인기계정</a>
-          </li>
-          <li>
-            <a href="#">해시태그</a>
-          </li>
-          <li>
-            <a href="#">위치</a>
-          </li>
-        </ul>
-        <span>© 2021 Instagram from Facebook</span>
-      </footer>
-    );
-  }
-}
-
-export default Footer;
