--- src/pages/summer/Main/Footer/Footer.scss
@@ -1,41 +1,0 @@
-@import '../../../../styles/common.scss';
-
-footer {
-  ul {
-    @include display-flex(unset, unset);
-    flex-wrap: wrap;
-    margin-bottom: 15px;
-
-    li {
-      position: relative;
-      padding: 3px 0;
-      color: $color-darkgray;
-      font-size: 12px;
-
-      &::after {
-        content: '';
-        position: absolute;
-        top: 50%;
-        right: -4px;
-        transform: translateY(-50%);
-        width: 1px;
-        height: 1px;
-        background: $color-darkgray;
-        border-radius: 50%;
-      }
-      & + li {
-        margin-left: 8px;
-      }
-      &:nth-child(7) {
-        margin-left: 0;
-      }
-      &:last-of-type::after {
-        background: transparent;
-      }
-    }
-  }
-  span {
-    color: $color-darkgray;
-    font-size: 12px;
-  }
-}
