+++ src/pages/summer/components/Footer/Footer.scss
@@ -1,0 +1,87 @@
+@import '../../../../styles/common.scss';
+
+div[class$='summer'] {
+  footer {
+    .footer-top {
+      flex-wrap: wrap;
+      margin-bottom: 17px;
+
+      li {
+        a {
+          color: $color-darkgray;
+          font-size: 12px;
+          text-decoration: none;
+        }
+      }
+    }
+
+    .footer-bottom {
+      span {
+        color: $color-darkgray;
+        font-size: 12px;
+      }
+    }
+  }
+}
+
+.Main-summer {
+  footer {
+    .footer-top {
+      @include display-flex(unset, unset);
+
+      li {
+        position: relative;
+        padding: 3px 0;
+
+        &::after {
+          content: '';
+          position: absolute;
+          top: 50%;
+          right: -4px;
+          width: 1.5px;
+          height: 1.5px;
+          background: $color-darkgray;
+          border-radius: 50%;
+        }
+        & + li {
+          margin-left: 8px;
+        }
+        &:nth-child(7) {
+          margin-left: 0;
+        }
+        &:last-of-type::after {
+          background: transparent;
+        }
+      }
+    }
+  }
+}
+
+.Login-summer {
+  footer {
+    .footer-top {
+      @include display-flex(center, unset);
+
+      li {
+        padding: 0px 2px;
+
+        & + li {
+          margin-left: 12px;
+        }
+      }
+    }
+
+    .footer-bottom {
+      @include display-flex(center, unset);
+
+      select {
+        margin-right: 10px;
+        width: 55px;
+        border: none;
+        background: none;
+        font-size: 12px;
+        color: $color-darkgray;
+      }
+    }
+  }
+}
